change sample coordinate to int

+++ app/src/main/java/jp/ac/oit/elc/mail/ibeaconlocationsystem/IntensitySample.java
@@ -7,12 +7,12 @@ import jp.ac.oit.elc.mail.ibeaconlocationsystem.wifi.WifiBeacon;
  * Created by yuuki on 10/6/15.
  */
 public class IntensitySample {
-    public float x;
-    public float y;
+    public int x;
+    public int y;
     private BeaconList<BluetoothBeacon> mBtBeaconList;
     private BeaconList<WifiBeacon> mWifiBeaconList;
 
-    public IntensitySample(float x, float y, BeaconList<BluetoothBeacon> btBeaconList, BeaconList<WifiBeacon> wifiBeaconList) {
+    public IntensitySample(int x, int y, BeaconList<BluetoothBeacon> btBeaconList, BeaconList<WifiBeacon> wifiBeaconList) {
         this.x = x;
         this.y = y;
         mBtBeaconList = btBeaconList;
